Moving API converters for tenant in dedicated file
