+++ packages/bff/src/model/domain/apiConverter.ts
@@ -4,7 +4,6 @@ import {
   attributeRegistryApi,
   authorizationApi,
   selfcareV2ClientApi,
-  tenantApi,
 } from "pagopa-interop-api-clients";
 import { P, match } from "ts-pattern";
 import { selfcareEntityNotFilled } from "./errors.js";
@@ -105,20 +104,4 @@ export const toApiAttributeProcessSeed = (
 ): attributeRegistryApi.CertifiedAttributeSeed => ({
   ...seed,
   code: createHash("sha256").update(seed.name).digest("hex"),
-});
-
-export const toBffApiCompactOrganization = (
-  input: tenantApi.Tenant
-): bffApi.CompactOrganization => ({
-  id: input.id,
-  name: input.name,
-});
-
-export const toBffApiRequesterCertifiedAttributes = (
-  input: tenantApi.CertifiedAttribute
-): bffApi.RequesterCertifiedAttribute => ({
-  tenantId: input.id,
-  tenantName: input.name,
-  attributeId: input.attributeId,
-  attributeName: input.attributeName,
 });
